Extract cloud backup reconciliation
Move the UniFFI reconcile callback types and reducer update helpers into a dedicated cloud backup reconciliation module while keeping the root manager exports stable.
