+++ rust/src/manager/cloud_backup_manager.rs
@@ -12,6 +12,7 @@ mod other_backups;
 mod pending;
 mod pending_enable;
 mod pending_verification;
+mod reconcile;
 mod remote_inventory;
 mod store;
 mod sync_health;
@@ -66,8 +67,7 @@ pub(crate) use self::error::{
 pub(crate) use self::keychain::CloudBackupKeychain;
 use self::model::{
     CloudBackupAcceptedEnablePrompt, CloudBackupExclusiveOperation,
-    CloudBackupExclusiveOperationClaim, CloudBackupStateReducer, CloudBackupStateReducerEffects,
-    CloudBackupStateReducerEvent,
+    CloudBackupExclusiveOperationClaim, CloudBackupStateReducer, CloudBackupStateReducerEvent,
 };
 pub use self::model::{CloudBackupLifecycle, CloudBackupRestoreFlow};
 pub(crate) use self::ops::{
@@ -86,12 +86,11 @@ pub(crate) use self::pending_enable::PendingEnableSessionMaterial;
 pub(crate) use self::pending_verification::{
     PendingVerificationCompletion, PendingVerificationUpload,
 };
+#[allow(unused_imports)]
+pub use self::reconcile::{CloudBackupManagerReconciler, CloudBackupReconcileMessage};
 pub(crate) use self::remote_inventory::current_namespace_wallet_record_ids;
 pub(crate) use self::store::CloudBackupStore;
 pub(crate) use self::sync_health::SYNC_HEALTH_MISSING_MASTER_KEY_MESSAGE;
-use self::verify::coordinator::{
-    CloudBackupVerificationCoordinator, CloudBackupVerificationEffect,
-};
 pub(crate) use self::wallet_changes::{LIVE_UPLOAD_DEBOUNCE, live_upload_retry_delay_for_attempt};
 #[cfg(test)]
 pub(crate) use self::wallets::UnpersistedPrfKey;
@@ -251,13 +250,6 @@ pub(crate) enum CloudBackupDisableOutcome {
     Started,
     ReturnedToIdle,
     Failed { message: String, can_keep_enabled: bool },
-}
-
-/// Typed state delta sent from Rust to Swift and Kotlin reconcilers
-#[derive(Debug, Clone, uniffi::Enum)]
-pub enum CloudBackupReconcileMessage {
-    Lifecycle(Box<CloudBackupLifecycle>, CloudBackupSettingsRowStatus),
-    EnableCompleted(CloudBackupEnableContext),
 }
 
 /// Restore summary shown after cloud backup onboarding restore completes
@@ -586,11 +578,6 @@ impl CloudBackupDetailResult {
     }
 }
 
-#[uniffi::export(callback_interface)]
-pub trait CloudBackupManagerReconciler: Send + Sync + std::fmt::Debug + 'static {
-    fn reconcile(&self, message: CloudBackupReconcileMessage);
-}
-
 #[derive(Clone, Debug, uniffi::Object)]
 pub struct RustCloudBackupManager {
     pub state: Arc<RwLock<CloudBackupStateReducer>>,
@@ -814,60 +801,6 @@ impl RustCloudBackupManager {
         manager
     }
 
-    fn load_persisted_flags() -> (CloudBackupVerificationMetadata, bool) {
-        let db_state = Self::load_persisted_state();
-        ((&db_state).into(), db_state.should_prompt_verification())
-    }
-
-    pub(super) fn send(&self, message: Message) {
-        if let Err(error) = self.reconciler.send(message) {
-            error!("unable to send cloud backup message: {error:?}");
-        }
-    }
-
-    fn apply_model_event(&self, event: CloudBackupStateReducerEvent) -> bool {
-        let effects = match self.state.write().apply_event(event) {
-            Ok(effects) => effects,
-            Err(rejection) => match rejection {},
-        };
-
-        self.send_model_effects(effects);
-        true
-    }
-
-    fn send_model_effects(&self, effects: CloudBackupStateReducerEffects) {
-        if let Some(lifecycle) = effects.lifecycle {
-            self.send(Message::Lifecycle(
-                Box::new(lifecycle.lifecycle),
-                lifecycle.settings_row_status,
-            ));
-        }
-
-        if let Some(context) = effects.enable_completed {
-            self.send(Message::EnableCompleted(context));
-        }
-    }
-
-    pub(crate) fn reconcile_runtime_status(&self, status: CloudBackupStatus) {
-        if !matches!(status, CloudBackupStatus::Enabled | CloudBackupStatus::Enabling) {
-            self.clear_runtime_passkey_authorization();
-        }
-
-        let event = CloudBackupStateReducerEvent::RuntimeStatusReconciled(status);
-        let effects = match self.state.write().apply_event(event) {
-            Ok(effects) => effects,
-            Err(rejection) => match rejection {},
-        };
-        let status_changed = effects.status_changed;
-        self.send_model_effects(effects);
-
-        if !status_changed {
-            return;
-        }
-
-        self.apply_model_event(CloudBackupStateReducerEvent::MissingPasskeyDismissalCleared);
-    }
-
     fn start_connectivity_listener(self: &Arc<Self>) {
         // use a weak reference so the listener thread exits when the manager is dropped
         let manager = Arc::downgrade(self);
@@ -918,53 +851,6 @@ impl RustCloudBackupManager {
                 send!(self.supervisor.start_verification(true));
             }
             None => {}
-        }
-    }
-
-    pub(crate) fn observe_sync_health(&self, sync_health: CloudSyncHealth) {
-        self.apply_model_event(CloudBackupStateReducerEvent::SyncHealthObserved(sync_health));
-    }
-
-    pub(crate) fn reconcile_verification_presentation(
-        &self,
-        presentation: CloudBackupVerificationPresentation,
-    ) {
-        self.apply_model_event(CloudBackupStateReducerEvent::VerificationPresentationReconciled(
-            presentation,
-        ));
-    }
-
-    pub(crate) fn current_verification_source(&self) -> CloudBackupVerificationSource {
-        CloudBackupVerificationCoordinator::current_source(
-            self.state.read().verification_presentation(),
-        )
-    }
-
-    pub(crate) fn apply_verification_effect(&self, effect: CloudBackupVerificationEffect) {
-        if let Some(detail) = effect.detail {
-            self.apply_detail_outcome(CloudBackupDetailOutcome::Refreshed(detail));
-        }
-
-        if let Some(pending_upload_verification) = effect.pending_upload_verification {
-            self.apply_pending_upload_verification_value(pending_upload_verification);
-        }
-
-        if let Some(presentation) = effect.presentation {
-            self.reconcile_verification_presentation(presentation);
-        }
-
-        if let Some(verification) = effect.verification {
-            self.apply_verification_outcome(CloudBackupVerificationOutcome::from_state(
-                verification,
-            ));
-        }
-
-        if let Some(recovery) = effect.recovery {
-            self.apply_recovery_outcome(CloudBackupRecoveryOutcome::from_state(recovery));
-        }
-
-        if effect.refresh_sync_health {
-            self.refresh_sync_health();
         }
     }
 
@@ -1067,79 +953,6 @@ impl RustCloudBackupManager {
 
     pub(crate) fn refresh_sync_health(&self) {
         send!(self.supervisor.request_sync_health_refresh());
-    }
-
-    pub(crate) fn refresh_persisted_flags(&self) {
-        let (verification_metadata, should_prompt_verification) = Self::load_persisted_flags();
-
-        self.apply_model_event(CloudBackupStateReducerEvent::VerificationFlagsReconciled {
-            metadata: verification_metadata,
-            should_prompt: should_prompt_verification,
-        });
-    }
-
-    fn apply_pending_upload_verification_value(&self, pending: PendingUploadVerificationState) {
-        self.apply_model_event(CloudBackupStateReducerEvent::PendingUploadVerificationReconciled(
-            pending,
-        ));
-    }
-
-    pub(crate) fn reconcile_pending_upload_verification(
-        &self,
-        pending: PendingUploadVerificationState,
-    ) {
-        self.reconcile_pending_upload_verification_for_source(
-            pending,
-            self.current_verification_source(),
-        );
-    }
-
-    pub(crate) fn reconcile_pending_upload_verification_for_source(
-        &self,
-        pending: PendingUploadVerificationState,
-        source: CloudBackupVerificationSource,
-    ) {
-        let (verification_metadata, should_prompt_verification) = Self::load_persisted_flags();
-        let event = CloudBackupStateReducerEvent::PendingUploadVerificationAndFlagsReconciled {
-            pending,
-            metadata: verification_metadata,
-            should_prompt: should_prompt_verification,
-        };
-        let effects = match self.state.write().apply_event(event) {
-            Ok(effects) => effects,
-            Err(rejection) => match rejection {},
-        };
-        let decision_pending = effects.verification_decision_pending;
-        let presentation_changed = effects.verification_presentation_changed;
-        self.send_model_effects(effects);
-
-        if presentation_changed || decision_pending {
-            return;
-        }
-
-        self.apply_verification_effect(CloudBackupVerificationCoordinator::pending_upload_state(
-            pending, source,
-        ));
-    }
-
-    pub(crate) fn refresh_pending_upload_verification_state(&self) {
-        self.reconcile_pending_upload_verification(
-            self.current_pending_upload_verification_state(),
-        );
-    }
-
-    pub(crate) fn current_pending_upload_verification_state(
-        &self,
-    ) -> PendingUploadVerificationState {
-        if self.has_pending_cloud_upload_verification() {
-            return PendingUploadVerificationState::Confirming;
-        }
-
-        if self.pending_verification_completion().is_some() {
-            return PendingUploadVerificationState::Confirming;
-        }
-
-        PendingUploadVerificationState::Idle
     }
 
     pub(crate) fn apply_verification_outcome(&self, outcome: CloudBackupVerificationOutcome) {
@@ -1348,16 +1161,6 @@ impl RustCloudBackupManager {
     #[uniffi::constructor]
     pub fn new() -> Arc<Self> {
         CLOUD_BACKUP_MANAGER.clone()
-    }
-
-    pub fn listen_for_updates(&self, reconciler: Box<dyn CloudBackupManagerReconciler>) {
-        let reconcile_receiver = self.reconcile_receiver.clone();
-
-        std::thread::spawn(move || {
-            while let Ok(field) = reconcile_receiver.recv() {
-                reconciler.reconcile(field);
-            }
-        });
     }
 
     pub fn state(&self) -> CloudBackupState {
